+++ frontend-client/app/(main)/contest/[contestId]/@tabs/problem/page.tsx
@@ -4,23 +4,19 @@ import type { ContestProblem } from '@/types/type'
 import { columns } from './_components/Columns'
 
 interface ContestProblemProps {
-  params: { id: string }
+  params: { contestId: string }
 }
 
 export default async function ContestProblem({ params }: ContestProblemProps) {
-  const { id } = params
+  const { contestId } = params
   const { problems }: { problems: ContestProblem[] } = await fetcher
     .get('problem', {
       searchParams: {
         take: 10,
-        contestId: id
+        contestId
       }
     })
     .json()
-
-  problems.forEach((problem) => {
-    problem.id = problem.problemId
-  })
 
   return (
     <DataTable
@@ -33,7 +29,7 @@ export default async function ContestProblem({ params }: ContestProblemProps) {
         submissionCount: 'w-[14%]',
         acceptedRate: 'w-[14%]'
       }}
-      name="problem"
+      linked
     />
   )
 }
