+++ frontend-client/app/(main)/contest/_components/FinishedContestTable.tsx
@@ -26,7 +26,7 @@ export default async function FinishedContestTable() {
           participants: 'w-1/5 md:w-1/6',
           status: 'w-1/4 md:w-1/6'
         }}
-        name="contest"
+        linked
       />
     </>
   )
