+++ frontend-client/app/problem/[problemId]/_components/DescriptionLoading.tsx
