+++ frontend-client/app/problem/[problemId]/layout.tsx
@@ -1,20 +1,20 @@
+import EditorResizablePanel from '@/components/EditorResizablePanel'
 import { fetcher } from '@/lib/utils'
 import codedangLogo from '@/public/codedang-editor.svg'
 import type { ProblemDetail } from '@/types/type'
 import Image from 'next/image'
 import Link from 'next/link'
 import { IoIosArrowForward } from 'react-icons/io'
-import MainResizablePanel from './_components/MainResizablePanel'
 
 export default async function layout({
   params,
   children
 }: {
-  params: { id: number }
+  params: { problemId: string }
   children: React.ReactNode
 }) {
-  const { id } = params
-  const data: ProblemDetail = await fetcher(`problem/${id}`).json()
+  const { problemId } = params
+  const problem: ProblemDetail = await fetcher(`problem/${problemId}`).json()
 
   return (
     <div className="grid-rows-editor grid h-dvh w-full min-w-[1000px] overflow-x-auto bg-slate-800 text-white">
@@ -26,11 +26,11 @@ export default async function layout({
           <div className="flex items-center gap-1">
             <Link href="/problem">Problem</Link>
             <IoIosArrowForward className="size-6" />
-            <h1 className="text-lg font-bold text-white">{`#${id}. ${data.title}`}</h1>
+            <h1 className="text-lg font-bold text-white">{`#${problemId}. ${problem.title}`}</h1>
           </div>
         </div>
       </header>
-      <MainResizablePanel data={data}>{children}</MainResizablePanel>
+      <EditorResizablePanel problem={problem}>{children}</EditorResizablePanel>
     </div>
   )
 }
