+++ frontend-client/app/problem/[problemId]/submission/_components/Columns.tsx
