+++ frontend-client/components/EditorDescription.tsx
@@ -5,26 +5,18 @@ import {
   AccordionTrigger
 } from '@/components/ui/accordion'
 import { Badge } from '@/components/ui/badge'
-import { fetcher } from '@/lib/utils'
 import type { ProblemDetail } from '@/types/type'
 import { sanitize } from 'isomorphic-dompurify'
 import { Lightbulb, Tag } from 'lucide-react'
 
-export default async function DescriptionPage({
-  params
-}: {
-  params: { id: number }
-}) {
-  const { id } = params
-  const data: ProblemDetail = await fetcher(`problem/${id}`).json()
-
+export function EditorDescription({ problem }: { problem: ProblemDetail }) {
   return (
     <div className="flex h-full flex-col gap-8 p-6 text-lg">
       <div>
-        <h1 className="mb-3 text-xl font-bold">{`#${data.id}. ${data.title}`}</h1>
+        <h1 className="mb-3 text-xl font-bold">{`#${problem.id}. ${problem.title}`}</h1>
         <div
           className="prose prose-invert max-w-full text-sm leading-relaxed text-slate-300"
-          dangerouslySetInnerHTML={{ __html: sanitize(data.description) }}
+          dangerouslySetInnerHTML={{ __html: sanitize(problem.description) }}
         />
       </div>
       <div>
@@ -32,7 +24,7 @@ export default async function DescriptionPage({
         <div
           className="prose prose-invert max-w-full text-sm leading-relaxed text-slate-300"
           dangerouslySetInnerHTML={{
-            __html: sanitize(data.inputDescription)
+            __html: sanitize(problem.inputDescription)
           }}
         />
       </div>
@@ -41,15 +33,15 @@ export default async function DescriptionPage({
         <div
           className="prose prose-invert max-w-full text-sm leading-relaxed text-slate-300"
           dangerouslySetInnerHTML={{
-            __html: sanitize(data.outputDescription)
+            __html: sanitize(problem.outputDescription)
           }}
         />
       </div>
       <div className="flex items-center gap-3 text-base">
         <h2>Time Limit:</h2>
-        <p className="text-slate-300">{data.timeLimit} ms</p>
+        <p className="text-slate-300">{problem.timeLimit} ms</p>
         <h2>Memory Limit:</h2>
-        <p className="text-slate-300">{data.memoryLimit} MB</p>
+        <p className="text-slate-300">{problem.memoryLimit} MB</p>
       </div>
       <Accordion type="multiple">
         <AccordionItem value="item-1" className="border-b-slate-700">
@@ -60,7 +52,7 @@ export default async function DescriptionPage({
             </div>
           </AccordionTrigger>
           <AccordionContent>
-            {data.tags.map((tag) => (
+            {problem.tags.map((tag) => (
               <Badge
                 key={tag.id}
                 className="bg-slate-300 text-slate-800 hover:bg-slate-300"
@@ -70,7 +62,7 @@ export default async function DescriptionPage({
             ))}
           </AccordionContent>
         </AccordionItem>
-        {data.hint && (
+        {problem.hint && (
           <AccordionItem value="item-2" className="border-b-slate-700">
             <AccordionTrigger className="hover:no-underline">
               <div className="flex items-center gap-2 text-base">
@@ -80,7 +72,7 @@ export default async function DescriptionPage({
             </AccordionTrigger>
             <AccordionContent>
               <div
-                dangerouslySetInnerHTML={{ __html: sanitize(data.hint) }}
+                dangerouslySetInnerHTML={{ __html: sanitize(problem.hint) }}
                 className="prose prose-invert max-w-full text-sm leading-relaxed text-slate-300"
               />
             </AccordionContent>
