+++ frontend-client/types/type.ts
@@ -64,7 +64,6 @@ export interface Standings {
 export interface ContestProblem {
   id: number
   order: number
-  problemId: number
   title: string
   difficulty: Level
   submissionCount: number
